[Update/#42]: Code Review 적용
